+++ src/main/java/cn/bmwm/modules/shop/dao/ProductDao.java
@@ -315,10 +315,10 @@ public interface ProductDao extends BaseDao<Product, Long> {
 	 * @param page：页码
 	 * @param size：每页显示商品数量
 	 * @param order：排序方式，1：推荐，2：人气，3：距离，4：价格
-	 * @param x：纬度
-	 * @param y：经度
+	 * @param x：经度
+	 * @param y：纬度
 	 * @return
 	 */
-	ItemPage<Product> findShopProductList(Shop shop, Integer type, ShopCategory category, Integer page, Integer size, Integer order, Integer x, Integer y);
+	ItemPage<Product> findShopProductList(Shop shop, Integer type, ShopCategory category, Integer page, Integer size, Integer order, Double x, Double y);
 
 }
